// src/interfaces/INotifications.ts
export interface INotifications {
  id_ticket: number;
  id_rastreador: number;
  dt_abertura: string;
  assunto: string;
  num_ticket: string;
  status: string;
  prioridade: string;
  tipo: string;
  chassi: string;
  placa: string;
  tipo_veiculo: string;
  id_cerca_virtual: number;
  nome_cerca: string;
}